requirements.txt
streamlit
transformers
google-generativeai
nltk